Simplify tests
Move the extra annotations test into tests/handlers/lab_test so
that all of the tests for the lab routes are together.  Strip
some unnecessary bits out of the test configurations so that they're
more minimal and it's more obvious what configuration options are
meaningful for the test they're running.
