+++ tests/data/extra-annotations/input/config.yaml
@@ -3,45 +3,6 @@ safir:
   logLevel: DEBUG
 lab:
   spawnTimeout: 10
-  env:
-    API_ROUTE: /api
-    AUTO_REPO_SPECS: "https://github.com/lsst-sqre/system-test@prod"
-    CULL_KERNEL_IDLE_TIMEOUT: "432000"
-    CULL_KERNEL_CONNECTED: "True"
-    CULL_KERNEL_INTERVAL: "300"
-    FIREFLY_ROUTE: /portal/app
-    HUB_ROUTE: /nb/hub
-    NO_ACTIVITY_TIMEOUT: "432000"
-    TAP_ROUTE: /api/tap
-
-    # Note that other environment variables will be injected by the Lab
-    # controller (to match KubeSpawner's behavior).  These must be calculated
-    # from other Lab-spawn-time values, and include:
-    #
-    # ACCESS_TOKEN
-    # CPU_GUARANTEE
-    # CPU_LIMIT
-    # DEBUG
-    # JPY_API_TOKEN
-    # JUPYTERHUB_ACTIVITY_URL
-    # JUPYTERHUB_CLIENT_ID
-    # JUPYTERHUB_OAUTH_ACCESS_SCOPES
-    # JUPYTERHUB_OAUTH_CALLBACK_URL
-    # JUPYTERHUB_OAUTH_SCOPES  # Alias for JUPYTERHUB_OAUTH_ACCESS_SCOPES
-    # JUPYTERHUB_SERVICE_PREFIX
-    # JUPYTERHUB_SERVICE_URL
-    # JUPYTERHUB_USER
-    # JUPYTER_IMAGE         # Alias for JUPYTER_IMAGE_SPEC
-    # JUPYTER_IMAGE_SPEC
-    # MEM_GUARANTEE  # in bytes
-    # MEM_LIMIT      # in bytes
-    # RESET_USER_ENV
-    #
-    # Additionally, some environment variables will be set from the secrets
-    # below.
-    #
-    # EXTERNAL_INSTANCE_URL is a special case: it's constant-per-instance, but
-    # you can't template values files
   extraAnnotations:
     k8s.v1.cni.cncf.io/networks: "kube-system/dds"
   files:
@@ -49,113 +10,14 @@ lab:
       modify: true
       contents: |
         root:x:0:0:root:/root:/bin/bash
-        bin:x:1:1:bin:/bin:/sbin/nologin
-        daemon:x:2:2:daemon:/sbin:/sbin/nologin
-        adm:x:3:4:adm:/var/adm:/sbin/nologin
-        lp:x:4:7:lp:/var/spool/lpd:/sbin/nologin
-        sync:x:5:0:sync:/sbin:/bin/sync
-        shutdown:x:6:0:shutdown:/sbin:/sbin/shutdown
-        halt:x:7:0:halt:/sbin:/sbin/halt
-        mail:x:8:12:mail:/var/spool/mail:/sbin/nologin
-        operator:x:11:0:operator:/root:/sbin/nologin
-        games:x:12:100:games:/usr/games:/sbin/nologin
-        ftp:x:14:50:FTP User:/var/ftp:/sbin/nologin
-        tss:x:59:59:Account used by the trousers package to sandbox the tcsd daemon:/dev/null:/sbin/nologin
-        dbus:x:81:81:System message bus:/:/sbin/nologin
-        nobody:x:99:99:Nobody:/:/sbin/nologin
-        systemd-network:x:192:192:systemd Network Management:/:/sbin/nologin
-        lsst_lcl:x:1000:1000::/home/lsst_lcl:/bin/bash
     /etc/group:
       modify: true
       contents: |
         root:x:0:
-        bin:x:1:
-        daemon:x:2:
-        sys:x:3:
-        adm:x:4:
-        tty:x:5:
-        disk:x:6:
-        lp:x:7:
-        mem:x:8:
-        kmem:x:9:
-        wheel:x:10:
-        cdrom:x:11:
-        mail:x:12:
-        man:x:15:
-        dialout:x:18:
-        floppy:x:19:
-        games:x:20:
-        utmp:x:22:
-        tape:x:33:
-        utempter:x:35:
-        video:x:39:
-        ftp:x:50:
-        lock:x:54:
-        tss:x:59:
-        audio:x:63:
-        dbus:x:81:
-        screen:x:84:
-        nobody:x:99:
-        users:x:100:
-        systemd-journal:x:190:
-        systemd-network:x:192:
-        cgred:x:997:
-        ssh_keys:x:998:
-        input:x:999:
-    /opt/lsst/software/jupyterlab/lsst_dask.yml:
-      contents: |
-        # No longer used, but preserves compatibility with runlab.sh
-        dask_worker.yml: |
-          enabled: false
-    /opt/lsst/software/jupyterlab/panda:
-      modify: false
-      contents: |
-        # Licensed under the Apache License, Version 2.0 (the "License");
-        # You may not use this file except in compliance with the License.
-        # You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
-        #
-        # Authors:
-        # - Wen Guan, <wen.guan@cern.ch>, 2020
-        [common]
-        # if logdir is configured, idds will write to idds.log in this directory.
-        # else idds will go to stdout/stderr.
-        # With supervisord, it's good to write to stdout/stderr, then supervisord can manage and rotate logs.
-        # logdir = /var/log/idds
-        loglevel = INFO
-        [rest]
-        host = https://iddsserver.cern.ch:443/idds
-        #url_prefix = /idds
-        #cacher_dir = /tmp
-        cacher_dir = /data/idds
-  initContainers:
-    - name: initdir
-      image: ghcr.io/lsst-sqre/initdir:0.0.3
-      privileged: true
-      volumes:
-        - containerPath: /home
-          mode: rw
-          source:
-            type: nfs
-            serverPath: /share1/home
-            server: 10.87.86.26
-  pullSecret: pull-secret
-  secrets:
-    - secretName: nublado-secret
-      secretKey: butler-secret
-      env: BUTLER_SECRET
-      path: /opt/lsst/software/jupyterlab/butler-secret
-    - secretName: extra-secret
-      secretKey: db-password
   sizes:
     small:
       cpu: 1.0
       memory: 3Gi
-    medium:
-      cpu: 2.0
-      memory: 6Gi
-    large:
-      cpu: 4.0
-      memory: 12Gi
   volumes:
     - containerPath: /home
       mode: rw
@@ -163,42 +25,8 @@ lab:
         type: nfs
         server: 10.13.105.122
         serverPath: /share1/home
-    - containerPath: /project
-      mode: ro
-      source:
-        type: hostPath
-        path: /share1/project
-    - containerPath: /scratch
-      mode: rw
-      source:
-        type: persistentVolumeClaim
-        storageClassName: sdf-home
-        accessModes:
-          - ReadWriteMany
-        resources:
-          requests:
-            storage: 1Gi
-    - containerPath: /temporary
-      subPath: temporary
-      mode: rw
-      source:
-        type: persistentVolumeClaim
-        storageClassName: sdf-home
-        accessModes:
-          - ReadWriteMany
-        resources:
-          requests:
-            storage: 1Gi
-fileserver:
-  image: docker.io/lsstsqre/worblehat
-  tag: ajt-dev
-  timeout: 86400
 images:
   source:
     type: docker
     registry: lighthouse.ceres
     repository: library/sketchbook
-  recommendedTag: recommended
-  numReleases: 1
-  numWeeklies: 2
-  numDailies: 3
